requirements.txt
requests
psycopg2-binary

SUPERSET-LOAD-EXAMPLE
SUPERSET-LOG-LEVEL